main tasks done

+++ controllers/contactsControllers.js
@@ -3,16 +3,21 @@ import * as contactsService from "../services/contactsServices.js";
 import HttpError from "../helpers/HttpError.js";
 import ctrlWrapper from "../decorators/ctrlWrapper.js";
 
-export const getAllContacts = async (_, res) => {
-  const result = await contactsService.listContacts();
+export const getAllContacts = async (req, res) => {
+  const { _id: owner } = req.user;
+  const { page = 1, limit = 3 } = req.query;
+
+  const skip = (page - 1) * limit;
+  const result = await contactsService.listContacts({ owner }, { skip, limit });
 
   res.json(result);
 };
 
 export const getOneContact = async (req, res) => {
   const { id } = req.params;
+  const { _id: owner } = req.user;
 
-  const result = await contactsService.getContactById(id);
+  const result = await contactsService.getContactById({ _id: id, owner });
 
   if (!result) {
     throw HttpError(404, `Contact with id - ${id} not found`);
@@ -23,7 +28,9 @@ export const getOneContact = async (req, res) => {
 
 export const deleteContact = async (req, res) => {
   const { id } = req.params;
-  const result = await contactsService.removeContact(id);
+  const { _id: owner } = req.user;
+
+  const result = await contactsService.removeContact({ _id: id, owner });
 
   if (!result) {
     throw HttpError(404, `Contact with id - ${id} not found`);
@@ -33,15 +40,21 @@ export const deleteContact = async (req, res) => {
 };
 
 export const createContact = async (req, res) => {
-  const result = await contactsService.addContact(req.body);
+  const { _id: owner } = req.user;
+
+  const result = await contactsService.addContact({ ...req.body, owner });
 
   res.status(201).json(result);
 };
 
 export const updateContact = async (req, res) => {
   const { id } = req.params;
+  const { _id: owner } = req.user;
 
-  const result = await contactsService.updateContactById(id, req.body);
+  const result = await contactsService.updateContactById(
+    { _id: id, owner },
+    req.body
+  );
 
   if (!result) {
     throw HttpError(404, `Contact with id - ${id} not found`);
@@ -52,10 +65,14 @@ export const updateContact = async (req, res) => {
 
 export const updateStatusContact = async (req, res) => {
   const { id } = req.params;
+  const { _id: owner } = req.user;
 
   const { favorite } = req.body;
 
-  const result = await contactsService.updateFavorite(id, favorite);
+  const result = await contactsService.updateFavorite(
+    { _id: id, owner },
+    favorite
+  );
 
   if (!result) {
     throw HttpError(404);
